// src/store/Calculator/types.ts
export enum Prices {
    standartMounth = 100,
    standartYear = 1000,
    advancedMounth = 150,
    advancedYear = 1400
}
export enum Currencies {
    RUB = 'RUB',
    CNY = 'CNY',
    KZT = 'KZT'
}

export enum Tarrifs {
    standart = 'standart',
    advanced = 'advanced'
}

export enum Periods {
    mounth = 'mounth',
    year = 'year'
}